requirements.txt
aws-cdk-lib
constructs

boto3~=1.34.152
aws-lambda-powertools~=2.43.1
numpy~=2.0.0
botocore~=1.34.152
pydexcom~=0.3.2